requirements.txt
# API runtime
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
anthropic>=0.49.0
httpx>=0.27.0
numpy>=1.26.0
pydantic>=2.6.0
python-dotenv>=1.0.0

# Embedding model (used by /api/extend at runtime)
sentence-transformers>=2.5.0
torch>=2.2.0
transformers>=4.38.0
